Separate server config from starting the server
